Update template to be postgres-only
