--- {{cookiecutter.bot_slug}}/Pipfile_gino
@@ -1,12 +1,0 @@
-[[source]]
-name = "pypi"
-url = "https://pypi.org/simple"
-verify_ssl = true
-
-[dev-packages]
-
-[packages]
-"discord.py" = "*"
-alembic = "*"
-gino = "*"
-psycopg2 = "*"
