+++ {{cookiecutter.bot_slug}}/config.json
@@ -1,4 +1,5 @@
 {
   "token": "{{cookiecutter.token}}",
-  "prefix": "{{cookiecutter.prefix}}"
+  "prefix": "{{cookiecutter.prefix}}",
+  "database": "postgresql://localhost:5432/bot"
 }
