--- {{cookiecutter.bot_slug}}/config_gino.json
@@ -1,5 +1,0 @@
-{
-  "token": "{{cookiecutter.token}}",
-  "prefix": "{{cookiecutter.prefix}}",
-  "database": "postgresql://localhost:5432/bot"
-}
